[TASK] Streamline `Build/Scripts/runTests.sh`
The `Build/Scripts/runTests.sh` dispatch wrapper
has been shifted away from the reality for this
extension and **must** be aligned, alone due to
the fact that it is used within the GitHub CI/CD
pipelines.

Note that it is really bad practice not covering
all PHP versions supported by the supported TYPO3
version, but not revertable `yet`. In the future
this should be respected for newer versions. This
is not company project and published for the wild.

* Remove PHP 7.4 and 8.0 support, not allowing to
  select them and use `PHP 8.1` as new default.
* Add `composer.json.*` pattern to `.gitignore`,
  used by `Build/Scripts/runTests.sh` to backup
  it during `-s composerUpdate`.
* GitHub images no longer provides the standalone
  `docker-compose` command in favour of the newer
  `docker compose` plugin. It's not reasonable to
  rework the `runTests.sh` script for now to deal
  with both variants and dropped at all, GitHub
  action workflow gets a new step to provide a
  intermedia solution to allow executing the other
  steps again.
* Use correct phpstan configuration and baseline
  file paths in deprecated `docker-compose.yml`.

In general `Build/Scripts/runTests.sh` should be
streamlined to the newer TYPO3 variant avoiding
`docker-compose` at all.