[FEATURE] Add configurable FlashMessage creation mode
Until now, flash message to communicate notification
has only been created when not redirecting to the
configured `listPid`, which has already been replaced
with redirecting based on a dedicated `redirectPageId`
configuration.

In the one or other project it may be still reasonable
to display the notification even when redirecting to
another page. This requires to take care to really
consume the extbase flashmessage queue, but is doable
with casual integration tools and not part of this
change but adding a dedicated `display messages` plugin
is planned to be added as a followup to ship at least
a working default.

This change introduces configurable FlashMessage creation
mode in three levels:

*  Plugin `NewJobForm` flexform setting
*  Fallback to TypoScript SETUP/CONSTANT setting,
   when the default `Use TypoScript` setting is
   configured in plugin flexform settings.
*  Pass flash message creation mode enum as part
   of the `AfterSaveJobEvent` to provide a way
   for deveopers to implement programatically
   way in projects to determine the setting as
   last resort.

On the PHP side the options are implemented as
PHP integer baked enum, which is possible due
to having PHP 8.1 as minimal PHP version set,
beside not beeing a good move but avoids the
need to use old-school constants here for now.

Following modes are possible:

* `FlashMessageCreationMode::SUPPRESS_WITH_CONFIGURED_REDIRECT_PAGE`
   (INT: 0) which is the default and matches the current
   existing behaviour to create flash messages only when
   redirecting to the new form page again.
*  `FlashMessageCreationMode::ALWAYS` (INT: 1) always create
   flash messages, even when redirecting to another page,
   which require to take care that flashmessages are really
   consumed. Dedicated consuming plugin will be added with
   a follow-up feature, but can be done with casual TYPO3
   integration patterns already.

Risk assesment is considerable `no risk` as the default
fallback reflects the current state to suppress flash
messages when redirecting to another page and no action
needed directly to change anything in projects.

That means that this feature is non-breaking and has
absolutly no impact in existing installation, until
configuration options are used to change the creation
behaviour.

Following fluid snippet can be used within a `USER_INT`
content element to consume and display the message
queue:

```xml
<f:flashMessages
  queueIdentifier="extbase.flashmessages.tx_academicjobs_newjobform"
/>
```

In general, this change on itself is not complete on
it's own, which is a known fact and will be handled.