[FEATURE] Add dedicated redirect page configuration for `NewJobForm`
Processing new job form transfers used the global extension setting
`listPid` to determine the action to take after successfully persist
a new job entry.

Until now, no flash messages are created and redirected to selected
`listPid`, otherwse flashmessage are created and redirected to the
newJobForm page and form display action.

Using the global `listPid` option is not really reasonable for this,
as projects may want to redirect to a dedicated page after persisting
new jobs which are not the `listPid` page. That means, this option
could not be used or reconfigured as it would break the `listPid`
based back link display within the detail plugin/action.

To dismantel the missuse of `listPid` a new configiguration is now
added in a defined multistep fallback layer:

*  NewJobForm flexform option `redirectPageId`: This option has the
   highest priority from all pure configuration based options.
*  `plugin.tx_academicjobs.settings.saveForm.fallbackRedirectPageId`
   as fallback when plugin settings is not set.
*  (Deprecated) Last fallback is to use the global `listPid` setting,
   which is deprecated and will be removed in upcoming versions.

Already existing `AfterSaveJobEvent` is extended:

* `getRequest()` recieving current request
* `getSettings()` to get current determined flexform and TypoScript
  settings.
* `getRedirectPageId(): ?int`: Current set page id to redirect,
  not reflecting deprecated listPid fallback.
* `setRedirectPageId(?int): self`: Can be used to set or remove
  the redirect page. Note, setting this to null will fallback
  to deprecated listPid redirect.

Deprecated listPid fallback usage will emit a `E_USER_DEPRECATED`
error and silently stop working in a upcoming version.

Risk assesment for this change can be defined as non-existing,
because the whole change is implemented in breaking-free manner.