[TASK] Enable static code analyzer
The static code analyzer PHPStan is now enabled
and executed as part of the GitHub workflow.

Current state and reported issues are added to
the baseline for now and due to the higher level
and the fact that not all packages has been green
on phpstan before.

Used command(s):

```shell
Build/Scripts/runTests.sh -t 11 -s composerUpdate \
&& Build/Scripts/runTests.sh -t 11 -s phpstanGenerateBaseline \
&& Build/Scripts/runTests.sh -t 12 -s composerUpdate \
&& Build/Scripts/runTests.sh -t 12 -s phpstanGenerateBaseline
```