[TASK] Add form validation
To enable server side validation of the newJobForm
the field settings needed to be added, as well as the
validator itself. While testing it was noticed that the
registry did not yet load the settings, so the
respective entry was made in the `Services.yaml`.
The controller actions required `Job` argument was
set to also possibly be `NULL` as reloading the page
after submitting the form would result in a TYPO3
error.