[TASK] Use LanguageServiceFactory instead of $TSFE
This change updates the `ImageUploadConverter` to no longer rely on
the global `$GLOBALS['TSFE']`` variable, which is being
deprecated[1].

The `LanguageServiceFactory` is now injected into the converter's
constructor and used to create a language service based on the
current request's language. This allows for proper localization of
error messages without needing to access the frontend controller
directly.

Additionally, fallback error messages were also improved for better
clarity when localization fails.

[1] https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/13.0/Breaking-102621-MostTSFEMembersMarkedInternalOrRead-only.html#breaking-102621-1701937690