[TASK] Streamline academic-jobs language files (*.xlf)
This change modernizes the translation file format for all .xlf files
in `EXT:academic_jobs` to align with current standards streamlined
for the rest of extensions[1].

Rebasing the current branch was a real pain with all .xlf files'
differences from different commits, so we decided to streamline .xlf
files just at the end with current labels.

Additionally, a typo `data-langague` has been fixed to
`data-language`.

[1] https://github.com/fgtclb/academic-extensions/commit/1fff8547ca97149958cdaebbf01fbc812f141cd8